+++ src/main/java/sayTheSpire/UIPatches/monsters/AbstractMonsterPatch.java
@@ -1,0 +1,8 @@
+import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+import sayTheSpire.events.TextEvent;
+import sayTheSpire.Output;
+
+public class AbstractMonsterPatch {
+
+}