+++ src/main/java/sayTheSpire/UIPatches/orbs/AbstractOrbPatch.java
@@ -1,14 +1,17 @@
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
 import com.megacrit.cardcrawl.orbs.AbstractOrb;
-import sayTheSpire.Output;
 import sayTheSpire.ui.elements.OrbElement;
+import sayTheSpire.Output;
 
-@SpirePatch(clz = AbstractOrb.class, method = "update")
 public class AbstractOrbPatch {
 
-    public static void Postfix(AbstractOrb __instance) {
-        if (__instance.hb.justHovered) {
-            Output.setUI(new OrbElement(__instance));
+    @SpirePatch(clz = AbstractOrb.class, method = "update")
+    public static class UpdatePatch {
+
+        public static void Postfix(AbstractOrb __instance) {
+            if (__instance.hb.justHovered) {
+                Output.setUI(new OrbElement(__instance));
+            }
         }
     }
 }